+++ jsonpath/__init__.py
@@ -13,6 +13,9 @@ from .exceptions import JSONPointerIndexError
 from .exceptions import JSONPointerKeyError
 from .exceptions import JSONPointerResolutionError
 from .exceptions import JSONPointerTypeError
+from .exceptions import RelativeJSONPointerError
+from .exceptions import RelativeJSONPointerIndexError
+from .exceptions import RelativeJSONPointerSyntaxError
 from .filter import UNDEFINED
 from .lex import Lexer
 from .match import JSONPathMatch
@@ -50,6 +53,9 @@ __all__ = (
     "match",
     "Parser",
     "RelativeJSONPointer",
+    "RelativeJSONPointerError",
+    "RelativeJSONPointerIndexError",
+    "RelativeJSONPointerSyntaxError",
     "resolve",
     "UNDEFINED",
 )